User 메인화면, 마이페이지, 찜목록 페이지 프론트 UI작업

+++ src/main/java/kr/co/solfood/owner/store/OwnerStoreService.java
@@ -10,7 +10,7 @@ public interface OwnerStoreService {
     int insertStore(OwnerStoreVO vo);
 
     // 점주 <-> 상점 조회
-    OwnerStoreVO getOwnerStore(int id);
+    OwnerStoreVO getOwnerStore(long id);
 
     // 상점 수정
     int updateStore(OwnerStoreVO vo);
