+++ src/main/java/kr/co/solfood/owner/store/OwnerStoreServiceImpl.java
@@ -25,7 +25,7 @@ public class OwnerStoreServiceImpl implements OwnerStoreService {
 
     // 점주 <-> 상점 조회
     @Override
-    public OwnerStoreVO getOwnerStore(int id) {
+    public OwnerStoreVO getOwnerStore(long id) {
         return ownerStoreMapper.selectStoreById(id);
     }
